Please enter the commit message for your changes. Lines starting

+++ models/dual_encoder.py
@@ -56,11 +56,11 @@ def dual_encoder_model(
     rnn_outputs, rnn_states = tf.nn.dynamic_rnn(
         cell,
         tf.concat(0, [context_embedded,
-                      float(context_token_len_avg),
-                      float(context_nums),
+                      tf.cast(context_token_len_avg,dtype = tf.float32),
+                      tf.cast(context_nums,dtype = tf.float32),
                       utterance_embedded,
-                      float(utterance_token_len_avg),
-                      float(utterance_nums)]),
+                      tf.cast(utterance_token_len_avg,dtype = tf.float32),
+                      tf.cast(utterance_nums, dtype = tf.float32)]),
         sequence_length=tf.concat(0, [context_len+2, utterance_len+2]),
         dtype=tf.float32)
     encoding_context, encoding_utterance = tf.split(0, 2, rnn_states.h)
